% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\docType{data}
\name{sampledataset}
\alias{sampledataset}
\title{Sample dataset}
\format{A data frame with 600 rows and 13 columns:
\itemize{
  \item{\eqn{y_1, y_2, y_3}}{ response variables}
  \item{\eqn{x_1, x_2, x_3, x_4, x_5, x_6, x_7, x_8, x_9, x_10}}{ predictors}
}}
\usage{
sampledataset
}
\description{
This is artificial data following changing multivariate liner structure.
}
\keyword{datasets}
